Added SV module for signed, fixed-point divider module

+++ asic/func_sim/include/Misc.hpp
@@ -18,7 +18,7 @@
 
 /*----- TYPEDEF -----*/
 // Both should the same number of fractional bits, the large_fp_t is merely to store intermediate results in computation
-using fix_pt_t = fpm::fixed</*base type*/ std::int16_t, /*intermediate type*/ std::int64_t, /*# fractional bits*/ 10>; // 6.10 fixed point
+using fix_pt_t = fpm::fixed</*base type*/ std::int32_t, /*intermediate type*/ std::int64_t, /*# fractional bits*/ 10>; // 6.10 fixed point
 using large_fp_t = fpm::fixed</*base type*/ std::int32_t, /*intermediate type*/ std::int64_t, /*# fractional bits*/ 10>; // 22.10 fixed point
 
 /*----- MACROS -----*/
